adds cleanup: adds enigma class, removes encrypt and decrypt classes, all tests passing
